Integrate Vaadin.Router into the app
