+++ app/index.js
@@ -80,8 +80,8 @@ module.exports = class VaadinElementsApplicationGenerator extends generators.Bas
     // Rename main element's file name
     if (this.properties.elementName !== 'vaadin-elements-app') {
       this.fs.move(
-        path.join(this.destinationPath(), 'src', 'vaadin-elements-app.js'),
-        path.join(this.destinationPath(), 'src', this.properties.elementName + '.js')
+        path.join(this.destinationPath(), 'src', 'app', 'vaadin-elements-app.js'),
+        path.join(this.destinationPath(), 'src', 'app', this.properties.elementName + '.js')
       );
     }
   }
