+++ app/templates/index.html
@@ -72,7 +72,7 @@ This program is available under Apache License Version 2.0, available at https:/
     <script src="./node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 
     <!-- Load your application shell -->
-    <script type="module" src="./src/<%= elementName %>.js"></script>
+    <script type="module" src="./src/app/<%= elementName %>.js"></script>
 
     <!-- Add any global styles for body, document, etc. -->
     <style>
