+++ app/templates/package.json
@@ -11,6 +11,7 @@
     "@polymer/iron-icon": "^3.0.0",
     "@polymer/iron-pages": "^3.0.0",
     "@polymer/iron-selector": "^3.0.0",
+    "@vaadin/router": "^1.1.0",
     "@vaadin/vaadin-core": "^12.0.0-alpha5",
     "@webcomponents/webcomponentsjs": "^2.1.1"
   },
