+++ app/templates/polymer.json
@@ -2,11 +2,12 @@
 {
   "root": ".",
   "entrypoint": "index.html",
-  "shell": "src/<%= elementName %>.js",
+  "shell": "src/app/<%= elementName %>.js",
   "fragments": [
-    "src/404.js",
-    "src/employee-list.js",
-    "src/employee-new.js"
+    "src/routes/config.js",
+    "src/views/404.js",
+    "src/views/employee-list.js",
+    "src/views/employee-new.js"
   ],
   "sources": [
     "src/**/*",
