+++ app/templates/src/routes/pages.js
@@ -1,0 +1,2 @@
+export const EMPLOYEE_LIST = '/employee-list';
+export const NEW_EMPLOYEE = '/employee-new';
